split support_functions.h into one header for each cpp file
This make builds faster as this way only required headers are included
in each compilation.

Before

	real	0m12,081s
	user	0m15,818s
	sys	0m1,847s

After

	real	0m11,835s
	user	0m14,357s
	sys	0m1,724s

Currently this is not a lot, but as the code will change, more headers
will be added, and if every one were added to support_functions.h, it
would slow down the compilation of all the files, as it is included in
each of them.

By the way, reorder the deader inclusions alphabetically in each file.

Later, the broadcasting related files could also be removed.